requirements.txt
ipykernel
beautifulsoup4
requests
tqdm
numpy
pandas
python-dotenv
boto3
